requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-allauth==0.40.0
django-bulma==0.6.0
django-crispy-forms==1.8.0
django-debug-toolbar==2.0
django-extensions==2.2.5
django-rest-swagger==2.1.2
djangorestframework==3.10.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
pathspec==0.6.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
simplejson==3.16.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
